Refactor: 대기열 만료 스케줄 추가
